+++ packages/browser-wallet/src/background/index.ts
@@ -4,10 +4,53 @@ import {
     InternalMessageType,
     MessageType,
 } from '@concordium/browser-wallet-message-hub';
-import { storedConnectedSites, storedSelectedAccount } from '@shared/storage/access';
+import { HttpProvider } from '@concordium/web-sdk';
+import { storedConnectedSites, storedSelectedAccount, storedJsonRpcUrl } from '@shared/storage/access';
 
+import JSONBig from 'json-bigint';
 import bgMessageHandler from './message-handler';
 import { forwardToPopup, HandleMessage, HandleResponse, RunCondition, setPopupSize } from './window-management';
+
+/**
+ * Determines whether the given url has been whitelisted by any account.
+ */
+async function isWhiteListedForAnyAccount(url: string): Promise<boolean> {
+    const urlOrigin = new URL(url).origin;
+    const connectedSites = await storedConnectedSites.get();
+    if (connectedSites) {
+        return Object.values(connectedSites).some((sites) => sites.includes(urlOrigin));
+    }
+    return false;
+}
+
+async function performRpcCall(
+    method: string,
+    params: string | undefined,
+    senderUrl: string,
+    onSuccess: (response: string | undefined) => void,
+    onFailure: (response: string) => void
+) {
+    const isWhiteListed = await isWhiteListedForAnyAccount(senderUrl);
+    if (isWhiteListed) {
+        const url = await storedJsonRpcUrl.get();
+        if (!url) {
+            onFailure('No JSON-RPC URL available');
+        } else {
+            const provider = new HttpProvider(url, fetch);
+            provider
+                .request(
+                    // We lose the method's typing when sending the message.
+                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+                    method as any,
+                    params && JSONBig.parse(params)
+                )
+                .then(onSuccess)
+                .catch((e) => onFailure(e.toString()));
+        }
+    } else {
+        onFailure('RPC Call can only be performed by whitelisted sites');
+    }
+}
 
 /**
  * Callback method which installs Injected script into Main world of Dapp
@@ -34,6 +77,17 @@ const injectScript: ExtensionMessageHandler = (_msg, sender, respond) => {
 bgMessageHandler.handleMessage(createMessageTypeFilter(InternalMessageType.Init), injectScript);
 bgMessageHandler.handleMessage(createMessageTypeFilter(InternalMessageType.SetViewSize), ({ payload }) => {
     setPopupSize(payload);
+});
+
+bgMessageHandler.handleMessage(createMessageTypeFilter(MessageType.JsonRpcRequest), (input, sender, respond) => {
+    const onFailure = (error: string) => respond({ success: false, error });
+    if (sender.url) {
+        const onSuccess = (response: string | undefined) => respond({ success: true, response });
+        performRpcCall(input.payload.method, input.payload.params, sender.url, onSuccess, onFailure);
+    } else {
+        onFailure('Missing sender URL');
+    }
+    return true;
 });
 
 /**